requirements.txt
aiosmtplib==2.0.2
alembic==1.13.0
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.1.1
blinker==1.7.0
cffi==1.16.0
click==8.1.7
cryptography==41.0.7
dnspython==2.4.2
email-validator==2.1.0.post1
exceptiongroup==1.1.3
fastapi==0.104.1
fastapi-mail==1.4.1
greenlet==3.0.1
h11==0.14.0
httptools==0.6.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
passlib==1.7.4
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==2.5.1
pydantic-settings==2.1.0
pydantic_core==2.14.3
PyJWT==2.8.0
pymongo==4.6.1
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
